Restructured File Tree
- put components in single file
- adjusted media queries in Example Sentences
- tweaked gitignore
