+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
 .env
 server.js
+notes.md
+input-questions.txt
