+++ notes.md
@@ -1,5 +1,3 @@
-//TODO -
-Find out why p tags get a pointer on hover by default.
 
 Error Msg in UserInput comp
