+++ src/index.js
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 
-import App from "./App";
+import App from "./components/App";
 
 const rootElement = document.getElementById("root");
 const root = createRoot(rootElement);
