+++ src/layouts/Layout.js
@@ -1,6 +1,6 @@
-import styles from "./styles/navbar.module.css";
+import styles from "../styles/navbar.module.css";
 import { Outlet } from "react-router-dom";
-import Navbar from "./Navbar";
+import Navbar from "../components/Navbar";
 
 export default function Layout() {
   return (
