+++ src/pages/DrillMaker.js
@@ -1,11 +1,11 @@
 //CONTEXT
 import { useContext } from "react";
-import { AppContext } from "../App";
+import { AppContext } from "../components/App";
 
 //COMPONENTS
-import UserInput from "../UserInput"
-import HighlightWords from "../HighlightWords"
-import ExampleSentences from "../ExampleSentences"
+import UserInput from "../components/UserInput"
+import HighlightWords from "../components/HighlightWords"
+import ExampleSentences from "../components/ExampleSentences"
 
 function DrillMaker() {
   const { state, dispatch } = useContext(AppContext);
