+++ src/routes/router.js
@@ -5,12 +5,12 @@ import {
   Route,
 } from "react-router-dom";
 //LAYOUT
-import Layout from "../Layout";
+import Layout from "../layouts/Layout";
 
 //PAGES
 import Homepage from "../pages/Homepage";
 import DrillMaker from "../pages/DrillMaker";
-import ErrorPage from "../ErrorPage";
+import ErrorPage from "../components/ErrorPage";
 
 export const router = createBrowserRouter(
   createRoutesFromElements(
