+++ src/styles/styles.css
@@ -1,0 +1,33 @@
+/* GLOBAL STYLES */
+
+* {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+  font-family: "Wix Madefor Display", sans-serif; 
+}
+
+a{
+  color: whitesmoke;
+  text-decoration: none;
+}
+
+.active-nav-link{
+  background-color: #2c3d55 ;
+}
+
+/* GOOGLE MATERIAL ICONS */
+.material-symbols-outlined {
+  font-variation-settings: "FILL" 0, "wght" 400, "GRAD" 0, "opsz" 48;
+  cursor: pointer;
+}
+
+/* WINDOW WIDTH STYLES */
+.winWidth{
+  position: absolute;
+}
+
+
+
+
+
